split up _plotting.py into multiple files
